broke down Header into more components
