+++ styles/components/header/HeaderDesktopMenu.module.sass
@@ -1,0 +1,4 @@
+@use '@/styles/utils/mixin' as m
+
+.hidden
+  display: none
