+++ styles/components/header/HeaderTabs.module.sass
@@ -1,0 +1,1 @@
+@use '@/styles/utils/mixin' as m
