+++ styles/components/layout/Header.module.sass
@@ -18,6 +18,3 @@
     display: table
   &::after
     clear: both
-
-.hidden
-  display: none
